# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/search": {
      "get": {
        "operationId": "SearchController_run",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["en", "ar"],
              "default": "en"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          },
          {
            "name": "retailer",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 80
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["relevance", "price", "unit_price"],
              "default": "relevance"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search products and compare prices across grocers (no account needed).",
        "tags": ["search"]
      }
    },
    "/v1/search/autocomplete": {
      "get": {
        "operationId": "SearchController_autocomplete",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 60
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10,
              "default": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["search"]
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["me"]
      },
      "patch": {
        "operationId": "MeController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "locale": {
                    "type": "string",
                    "enum": ["en", "ar"]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["me"]
      },
      "delete": {
        "operationId": "MeController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  }
                },
                "required": ["password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Erase the account (anonymises the row; removes personal content).",
        "tags": ["me"]
      }
    },
    "/v1/me/consents": {
      "patch": {
        "operationId": "MeController_setConsents",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "consents": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "purpose": {
                          "type": "string",
                          "enum": [
                            "account_terms",
                            "history",
                            "alerts_email",
                            "contribute_receipts",
                            "analytics",
                            "area_sync"
                          ]
                        },
                        "granted": {
                          "type": "boolean"
                        }
                      },
                      "required": ["purpose", "granted"],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 6
                  },
                  "version": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 20,
                    "default": "v0.1"
                  }
                },
                "required": ["consents"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Grant/withdraw consents. Each change appends to the consent ledger.",
        "tags": ["me"]
      }
    },
...
    "/v1/alerts": {
      "get": {
        "operationId": "AlertsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["alerts"]
      },
      "post": {
        "operationId": "AlertsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "threshold_qar": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "maximum": 5000
                  }
                },
                "required": ["product_id", "threshold_qar"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a price alert. Needs the alerts_email consent (PATCH /me/consents).",
        "tags": ["alerts"]
      }
    },
...
    "/v1/alerts/unsubscribe": {
      "post": {
        "operationId": "AlertsController_unsubscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 600
                  }
                },
                "required": ["token"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["alerts"]
      }
    },
...
  },
  "info": {
    "title": "Qarib API",
    "description": "Qatar grocery price comparison. Prices come from retailers, partners and community reports and may change.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qarib_at"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}